Refactor Reader integration into the HTTP client implementation
