+++ matchbook-sdk-rest/src/main/java/com/matchbook/sdk/rest/configs/wrappers/HttpClientWrapper.java
@@ -1,5 +1,6 @@
 package com.matchbook.sdk.rest.configs.wrappers;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -7,13 +8,12 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
+import com.matchbook.sdk.core.exceptions.ErrorType;
 import com.matchbook.sdk.core.exceptions.MatchbookSDKHttpException;
 import com.matchbook.sdk.rest.HttpConfig;
 import com.matchbook.sdk.rest.configs.HttpCallback;
-import com.matchbook.sdk.rest.configs.HttpCallback2;
 import com.matchbook.sdk.rest.configs.HttpClient;
-import com.matchbook.sdk.rest.dtos.Reader;
-import com.matchbook.sdk.rest.dtos.RestResponse;
+import okhttp3.Call;
 import okhttp3.Callback;
 import okhttp3.Cookie;
 import okhttp3.CookieJar;
@@ -22,6 +22,8 @@ import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
+import okhttp3.Response;
+import okhttp3.ResponseBody;
 
 public class HttpClientWrapper implements HttpClient {
 
@@ -88,89 +90,35 @@ public class HttpClientWrapper implements HttpClient {
     }
 
     @Override
-    public <T, R extends RestResponse<T>> void get(String url,
-            HttpCallback2<T, R> httpCallback,
-            Reader<T, R> reader) throws MatchbookSDKHttpException {
-        Request request = buildGetRequest(url);
-        sendHttpRequest(request, new RequestCallback2<>(httpCallback, reader));
-    }
-
-    @Deprecated
-    @Override
     public void get(String url, HttpCallback httpCallback) throws MatchbookSDKHttpException {
-        Request request = buildGetRequest(url);
-        sendHttpRequest(request, new RequestCallback(httpCallback));
-    }
-
-    @Override
-    public <T, R extends RestResponse<T>> void post(String url,
-            String body,
-            HttpCallback2<T, R> httpCallback,
-            Reader<T, R> reader) throws MatchbookSDKHttpException {
-        Request request = buildPostRequest(url, body);
-        sendHttpRequest(request, new RequestCallback2<>(httpCallback, reader));
-    }
-
-    @Deprecated
-    @Override
-    public void post(String url, String body, HttpCallback httpCallback) throws MatchbookSDKHttpException {
-        Request request = buildPostRequest(url, body);
-        sendHttpRequest(request, new RequestCallback(httpCallback));
-    }
-
-    @Override
-    public <T, R extends RestResponse<T>> void put(String url,
-            String body,
-            HttpCallback2<T, R> httpCallback,
-            Reader<T, R> reader) throws MatchbookSDKHttpException {
-        Request request = buildPutRequest(url, body);
-        sendHttpRequest(request, new RequestCallback2<>(httpCallback, reader));
-    }
-
-    @Deprecated
-    @Override
-    public void put(String url, String body, HttpCallback httpCallback) throws MatchbookSDKHttpException {
-        Request request = buildPutRequest(url, body);
-        sendHttpRequest(request, new RequestCallback(httpCallback));
-    }
-
-    @Override
-    public <T, R extends RestResponse<T>> void delete(String url,
-            HttpCallback2<T, R> httpCallback,
-            Reader<T, R> reader) throws MatchbookSDKHttpException {
-        Request request = buildDeleteRequest(url);
-        sendHttpRequest(request, new RequestCallback2<>(httpCallback, reader));
-    }
-
-    @Deprecated
-    @Override
-    public void delete(String url, HttpCallback httpCallback) throws MatchbookSDKHttpException {
-        Request request = buildDeleteRequest(url);
-        sendHttpRequest(request, new RequestCallback(httpCallback));
-    }
-
-    private Request buildGetRequest(String url) {
-        return buildRequest(url)
+        Request request = buildRequest(url)
                 .get()
                 .build();
+        sendHttpRequest(request, new RequestCallback(httpCallback));
     }
 
-    private Request buildPostRequest(String url, String body) {
-        return buildRequest(url)
+    @Override
+    public void post(String url, String body, HttpCallback httpCallback) throws MatchbookSDKHttpException {
+        Request request = buildRequest(url)
                 .post(RequestBody.create(body, jsonMediaType))
                 .build();
+        sendHttpRequest(request, new RequestCallback(httpCallback));
     }
 
-    private Request buildPutRequest(String url, String body) {
-        return buildRequest(url)
+    @Override
+    public void put(String url, String body, HttpCallback httpCallback) throws MatchbookSDKHttpException {
+        Request request = buildRequest(url)
                 .put(RequestBody.create(body, jsonMediaType))
                 .build();
+        sendHttpRequest(request, new RequestCallback(httpCallback));
     }
 
-    private Request buildDeleteRequest(String url) {
-        return buildRequest(url)
+    @Override
+    public void delete(String url, HttpCallback httpCallback) throws MatchbookSDKHttpException {
+        Request request = buildRequest(url)
                 .delete()
                 .build();
+        sendHttpRequest(request, new RequestCallback(httpCallback));
     }
 
     private Request.Builder buildRequest(String url) {
@@ -182,6 +130,60 @@ public class HttpClientWrapper implements HttpClient {
 
     private void sendHttpRequest(Request request, Callback callback) {
         httpClient.newCall(request).enqueue(callback);
+    }
+
+    private static class RequestCallback implements Callback {
+
+        private final HttpCallback httpCallback;
+
+        RequestCallback(HttpCallback httpCallback) {
+            this.httpCallback = httpCallback;
+        }
+
+        @Override
+        public void onResponse(Call call, Response response) {
+            try (ResponseBody responseBody = response.body()) {
+                if (response.isSuccessful() && Objects.nonNull(responseBody)) {
+                    httpCallback.onResponse(responseBody.byteStream());
+                } else {
+                    MatchbookSDKHttpException matchbookException = getExceptionForResponse(response);
+                    httpCallback.onFailure(matchbookException);
+                }
+            }
+        }
+
+        @Override
+        public void onFailure(Call call, IOException exception) {
+            MatchbookSDKHttpException matchbookException = new MatchbookSDKHttpException(exception.getMessage(), exception);
+            httpCallback.onFailure(matchbookException);
+        }
+
+        private MatchbookSDKHttpException getExceptionForResponse(Response response) {
+            ResponseBody responseBody = response.body();
+            if (Objects.nonNull(responseBody)) {
+                try {
+                    if (isAuthenticationErrorPresent(responseBody)) {
+                        return newUnauthenticatedException();
+                    }
+                } catch (IOException e) {
+                    return newHTTPException(response.code());
+                }
+            }
+            return newHTTPException(response.code());
+        }
+
+        private boolean isAuthenticationErrorPresent(ResponseBody body) throws IOException {
+            return new String(body.bytes()).toLowerCase().contains("cannot login");
+        }
+
+        private MatchbookSDKHttpException newHTTPException(int responseCode) {
+            return new MatchbookSDKHttpException("Unexpected HTTP code " + responseCode, ErrorType.HTTP);
+        }
+
+        private MatchbookSDKHttpException newUnauthenticatedException() {
+            return new MatchbookSDKHttpException("Incorrect username or password", ErrorType.UNAUTHENTICATED);
+        }
+
     }
 
 }
